+++ lib/options/highcharts_color_axis_options.dart
@@ -457,6 +457,9 @@ class HighchartsColorAxisOptions extends HighchartsOptionsBase {
   double? range;
 
   /// Whether to reverse the axis so that the highest number is closest
+  /// to the origin. Defaults to `false`.
+  ///
+  /// Whether to reverse the axis so that the highest number is closest
   /// to the origin. Defaults to `false` in a horizontal legend and
   /// `true` in a vertical legend, where the smallest value starts on
   /// top.